stock_ticker/requirements.txt
yfinance
pandas
numpy
ta
nltk
scikit-learn
tensorflow
streamlit
schedule
requests
feedparser
pyyaml
plotly
tqdm
nselib
